Add DaysWritable
